Fix absolutepath and introduce parentpath function in module. Fix some type emission. Maybe fix windows pdb files.

+++ src/common/source_location.cpp
@@ -24,11 +24,11 @@ namespace klong {
         return _source->code();
     }
 
-    uint64_t SourceLocation::column() const {
+    uint32_t SourceLocation::column() const {
         return _column;
     }
 
-    uint64_t SourceLocation::line() const {
+    uint32_t SourceLocation::line() const {
         return _line;
     }
